ad_reach/monitors/serializers.py
from rest_framework import serializers
from .models import Monitor, Slot

class MonitorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Monitor
        fields = '__all__'
        read_only_fields = ('vendor',)

class SlotSerializer(serializers.ModelSerializer):
    class Meta:
        model = Slot
        fields = '__all__'
