ad_reach/requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
celery
redis
channels
daphne
drf-spectacular
djangorestframework-simplejwt
sentry-sdk
django-cors-headers
dj-database-url
whitenoise
